+++ SourceryTests/Parsing/FileParser + MethodsSpec.swift
@@ -20,7 +20,7 @@ class FileParserMethodsSpec: QuickSpec {
                 }
 
                 it("extracts methods properly") {
-                    expect(parse("class Foo { func bar(some: Int) throws ->Bar {}; func foo() ->    Foo {}; func fooBar() rethrows {}; func fooVoid() {}; func fooInOut(some: Int, anotherSome: inout String) {} }")).to(equal([
+                    expect(parse("class Foo { func bar(some: Int) throws ->Bar {}; func foo() ->    Foo {}; func fooBar() rethrows {}; func fooVoid() {}; func fooInOut(some: Int, anotherSome: inout String) {} deinit {} }")).to(equal([
                         Class(name: "Foo", methods: [
                             Method(name: "bar(some: Int)", selectorName: "bar(some:)", parameters: [
                                 MethodParameter(name: "some", typeName: TypeName("Int"))
@@ -31,7 +31,8 @@ class FileParserMethodsSpec: QuickSpec {
                             Method(name: "fooInOut(some: Int, anotherSome: inout String)", selectorName: "fooInOut(some:anotherSome:)", parameters: [
                                 MethodParameter(name: "some", typeName: TypeName("Int")),
                                 MethodParameter(name: "anotherSome", typeName: TypeName("inout String"), `inout`: true)
-                                ], returnTypeName: TypeName("Void"), definedInTypeName: TypeName("Foo"))
+                                ], returnTypeName: TypeName("Void"), definedInTypeName: TypeName("Foo")),
+                            Method(name: "deinit", selectorName: "deinit", definedInTypeName: TypeName("Foo"))
                             ])
                         ]))
                 }