+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### New Features
 
 - Added support for file paths in `config` parameter
+- Added `isDeinitializer` property for methods
 - Improved config file validation and error reporting
 
 ### Bug fixes